requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain-chroma
langchain-core
langchain-text-splitters
chromadb
python-dotenv
tqdm
python-docx
langchain-google-genai
langchain-classic
sentence-transformers
rank_bm25
slowapi
langserve[all]
pydantic
